test: move fixtures, add e2e test for tracing and config
