+++ test/fixtures/config-module/config-module.config.mock.ts
