+++ test/fixtures/config-module/config-module.mock.ts
